backend/requirements.txt
Django==3.2.3
django-filter==23.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
djoser==2.1.0
Pillow==9.0.0